CMake: Handle QT_QML_SKIP_QMLDIR_ENTRY property
Add a new source file property QT_QML_SKIP_QMLDIR_ENTRY which tells
the build system not to add the source file as an entry in the
generated qmldir file.

This is useful to mark a file for installation via
qt6_target_qml_files without exposing it as a QML
type (like a helper javascript file).

Change-Id: I2aa595692bcbb65925635abc1b12519d1a45fe53
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>