qqml.h: do not use QML_REGISTER_TYPE(QObject)
This breaks code if the following circumstances all apply:

- A project is compiled with CMake using AUTOMOC
- It contains a class with a property of type QQmlList<QObject>
- In the resulting mocs_compilation.cpp, qqml.h is (transitively
  included).

Then, a use of QMetaTypeId<QQmlList<QObject>> will be in a
single file before its definition. The use comes from the instation
of qt_metaTypeArray. The explicit specialization comes from QMetaTypeID,
and happens afterwards (depending on include order).
gcc does not like this, and the compilation fails. There is probably no
reason for QML_REGISTER_TYPE(QObject), so we just remove it.

Removes the Q_DECLARE_METATYPE(QVariant), too, while we're at it.

Change-Id: Ia6607c31af92dadd809633188bb82fa90efaf922
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>