tst_qquickheaderview: wait for window to be active
The tst_QQuickHeaderView::listModel test fails on Ubuntu. From inspecting
the build log, VerticalHeaderView has an empty size. Since HeaderView
will only load items that end up visible in the view, it will load just one
item in this case, even if the model contains five. Hence the test will fail
when searching for the other missing items.

Why VerticalTableView has an empty size is not well understood (and cannot
be reproduced on my mac). The size is bound to the size of the window, which
indicates that the window is not shown or ready by the time the test runs.

So this patch will change how we create the window for the test, and
wait for it to be active, which is standard procedure for many other
tests. Hopefully this will also fix the test failure on Ubuntu.

Task-number: QTBUG-83172
Change-Id: I7d3a188dbeb161f11eee7b95c103928404161f11
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>