Text: Fix kerning when horizontalAlignment is set to center
When horizontalAlignment is set to center, alignedX will potentially
return a position that is not pixel-aligned. This causes issues with
text layout lower in the stack, leading to incorrect kerning.
Rounding the position makes sure we always have a pixel-aligned
position.

Note that this adapts the same alignment approach as centering done
by anchors, so the two will now be more consistent.

[ChangeLog][QtQuick][Text] Fixed kerning when horizontalAlignment is
set to center.

Fixes: QTBUG-49646
Change-Id: I6ce660b0c12435a27998cba78f2c0e98d1d92cc6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>