Move Default style out into its own plugin
In upcoming patches, we start registering C++ types declaratively.
A condition of doing so requires that each .pro corresponds to one
QML module. This conflicts with the QtQuick.Controls import, which
currently does quite a lot:

- Registers (and selects) QML files for the style that was set
- Registers private C++ utility types (such as IconLabel) that are
  useful for all styles under the QtQuick.Controls.impl import
- Registers private C++ types that are only useful for the Default
  style (such as BusyIndicatorImpl).

The reason it does so much can probably be explained by the
intended usage of Qt Quick Controls 2; when you do

    import QtQuick.Controls 2.0

you get access to the QML types (e.g. Button) that the style
you're using provides. So if you're using the Material style,
you'll get a Material style button. API-wise, the button is
identical to any other button, because the types in
QtQuick.Templates are what we advertise as the public API.

If we didn't have this functionality, users would need to
import specific style imports to use controls, and the
convenience of being able to simply start the application
with a different style by e.g. passing an application argument
would be lost.

To support declarative registration of types while also supporting
the existing use cases, we split out the Default-style-specific
stuff into a QtQuick.Controls.Default import.

Task-number: QTBUG-82922
Change-Id: Ib4f1620cae78d7acdc13d9ac0752a020bc22f3ea
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>