qmltyperegistrar: Follow the private includes setting in qmltypes
If the file names need to be prepended with "private/" in order to be
usable, that also holds for other tools that want to use them.
Therefore, also add the prefix to the qmltypes files. We won't have a
chance to figure this out again at a later point.

Furthermore, don't add file entries for types we cannot register anyway.

Change-Id: If04297db419e364f472b8ec627c599c0cbb4ffd9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>