CMake: Exclude building apps when cross-compiling
It would fail the iOS multi-arch build otherwise, at link time
with

Undefined symbols for architecture x86_64:
  "_adler32", referenced from:
      _png_icc_set_sRGB in libQt6BundledLibpng.a(png.c.o)<Paste>
....

This is due to us using system zlib instead of bundled zlib. Because
system zlib only has one slice of architectures (either device or
simulator), we get the above linker error.

Solution is not to link (build) the apps at all).

This is consistent with what qmake does.

Change-Id: Ief2e63a8da49f9daac4f5542e78b0f0b3d37ed93
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>