qmldom: Start of the Qml Dom library
The goal of the Dom library is to provide a nicer to use basis for the
Qml Code model, to be used by the various QML tools, the designer and
the new compiler.

This commit just adds some common utilities used by the library, to
output strings to a Sink, i.e. a function<void(QStringView).
This allows writing without allocations of extra temporary strings.
The more interesting parts are added in subsequent commits

Task-number: QTBUG-74840
Change-Id: I8fa43d5b622f59c8761b2469551127c0508c23c3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>