Undeprecate (in docs) QSGSimpleRect/TextureNode
In 5.8 the deprecated tag was added since these are not functional when
using a scenegraph backend other than default or software. In 6.0 we no
longer have the d3d12 backend so out of the box QSGSimpleRectNode and
QSGSimpleTextureNode are functional always.

The factory-style QQuickWindow::createImageNode() alternatives are still
preferred, but QSGSimpleTextureNode is used in a number of examples by
subclassing, and the factory function approach is no alternative to
that. Same goes internally, QQuickFramebufferObject is internally
enabled by subclassing QSGSimpleTextureNode. By extension, it is assumed
that there are applications out there relying on these classes too.

So solve the dilemma by removing the \deprecated from the docs and
that's that.

Task-number: QTBUG-82997
Change-Id: I30e8ee6df91fa37a370bba7a0511c95275794e33
Reviewed-by: Andy Nichols <andy.nichols@qt.io>