Fix clipping of text with software backend
QRawFont::boundingRect() returns the bounding rect of the glyph
as reported by the system. On Windows, this unfortunately
does not always include the antialiasing on the edges when
the ClearType renderer is used (probably for compatibility
reasons, since it matches the rendered size of the 8-bit alphamap.)

Therefore, we have traditionally been applying a margin around
rendering for 32-bit alpha maps, since there is no way to know
for certain the exact rendered size with ClearType.

It is tempting to expose the alphaMapBoundingBox through
QRawFont (either replace the implementation of boundingRect()
or add a new function) to avoid using private APIs here, but since
it is a work-around and we hope the need for it will go away when
we start using DirectWrite instead, it is better to keep it
locked away for now. It is also too late to do such a change in
Qt 5.15, so we can revisit the idea for Qt 6.x.

[ChangeLog][Software Renderer] Fixed a bug where the last column of
pixels in text would sometimes be clipped away.

Fixes: QTBUG-80180
Change-Id: Ic611a80aef3e0e9b9cee0f3a22c9e24f717d5dfe
Reviewed-by: Andy Nichols <andy.nichols@qt.io>