Add _q_dumpLayoutTree() for debugging purposes
I found this quite useful for debugging layouts in order to get a snapshot of
the layout hierarchy at any point. It will show the hierarchy and the relevant
properties that affect the layouting. It's basically QObject::dumpObjectTree()
specialized for Qt Quick Layouts (except it will dump it as QML-ish).
For the leaf nodes it will leave out unspecified information.

Example on how a dump looks like:

RowLayout {
    // Effective calculated values:
    sizeHintDirty: 1
    sizeHint.min : [  100,    20]
    sizeHint.pref: [  100,    20]
    sizeHint.max : [  inf,    20]

    RowLayout {
        // Effective calculated values:
        sizeHintDirty: 1
        sizeHint.min : [  100,    20]
        sizeHint.pref: [  100,    20]
        sizeHint.max : [  inf,    20]

        Rectangle {
            implicitWidth: 100
            implicitHeight: 20
        }

        Rectangle {
            implicitHeight: 20
            Layout.fillWidth: true
        }
    }
}

Pick-to: 5.15
Change-Id: Ie2fcd56d522395cb9d6f55a1ea3f882a7a9da709
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>