Allow loading images from arbitrary QIODevice subclasses
For PDF we need to pass a document object into QPdfIOHandler, but
QImageIOHandler only takes a QIODevice. A subclass of QFile is one way
to pass that extra object along.

So the plan is that a private PDF document wrapper class can inherit
QFile; and there will be a subclass of QQuickImage called PdfPageImage,
which will override QQuickImage::load() to construct the wrapper, which
will then get sent as QIODevice* into QPdfIOHandler. It will attempt to
cast that to the known private class; if it succeeds, it already will
have the FPDF_Document and will therefore not need to construct one.

Task-number: QTBUG-77506
Task-number: QTBUG-83988
Task-number: QTBUG-96574
Change-Id: I19c90a9f2c7060574720f831b3394de6c3714383
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>