rhi: Make rendercontext query safe in slots connected to invalidated
separateIndexBuffers() returns an incorrect value when called from a
slot connected to the invalidated() signal. This is because members like
m_rhi are nulled out before emitting the signal (thus making isValid()
return false), and this follows the existing QOpenGLContext (m_gl) behavior.
Instead of altering the logic, just decide the value of separateIndexBuffers
in initialize() and keep returning that from the query from then on.

This is relevant for Qt Quick 3D at least, which happens to destroy manually
created QSGLayer instances (which in turn own renderer objects) in a slot
connected to invalidated().

Task-number: QTBUG-83469
Change-Id: I81b34b57921242367c80f7d04331cb6bd3318ec8
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>