shapes: Avoid upsetting Mesa on eglfs with kms
We need to either ensure the context is destroyed before the offscreen
surface, or make sure the context is not current (and so the surface is
not associated with it anymore) when it is destroyed. We choose the
latter here.

Change-Id: If4424b98c4d8a718aa49d99285360ded6a5e128e
Fixes: QTBUG-82371
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 239a90b8d5588bc6536cb81bbba7af8cc049d945)
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>