Fix using fonts inside .ttc with render type QtRendering
The file name of a font is not identification in itself, because
the file may be a font collection (.ttc). In this case, we also
have to use the index of the font inside the .ttc as part of
the cache key.

[ChangeLog][QtQuick][Text] Fixed a bug where we would sometimes
resolve the wrong fonts inside .ttc font collections.

Fixes: QTBUG-81810
Change-Id: I40ded703bcdd22dce2be58c5a481f0918e6d352b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 763897702dfc22075a0366df8dfd8f381103a60b)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>