QQuickWindow: pass InputMethod(Query) events on to items
The events were not handled at all on window level, so input from e.g
the macOS Emoji keyboard never arrived in the Qt Quick items.

Since the input method is likely to use a separate window, the
QQuickWindow is likely to not have an active focus item when the input
method event arrives. In order to deliver the event to the correct item,
walk the scopedFocusItem chain starting with the contentItem as a fall
back.

[ChangeLog][QtQuick][QtQuickWindow] Pass input method events on
to the focus item.

Change-Id: I452bb73e13f9cd398c2302303f52a92bda2bb0e3
Fixes: QTBUG-70319
Pick-to: 5.15
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>