Make no-opengl builds succeed with Vulkan or Metal
Amends c83ca003248a21e368ebfbf799ef9580e0f4e6ac.

In Qt 5.x we do not support using Qt Quick with the QRhi-based
rendering path in -no-opengl builds. (such builds default to the
software backend in fact)

This is due to structural difficulties, such as QSGDefaultRenderContext
(the .cpp) being tied to direct OpenGL. This will only get solved properly
in Qt 6 when the direct OpenGL code path gets removed.

However, compilation should not break in -no-opengl builds,
regardless of which other conditions (QT_CONFIG(vulkan), Q_OS_MACOS,
etc.) evaluate to true.

The original patch missed the case when we have Vulkan enabled
or we are on macOS or iOS (which implies enabling the Metal code
path). This is now corrected by extending the scope of the condition.

Change-Id: I0e87d8b033aed8fdfaca826bcd32596a532380a7
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>