Fix font-related test failures triggered by Qt Base update
Change f761ad3cd9ad1252f24b76ae413298dc7bed8af3 in qtbase exposes
a pre-existing difference in measuring the height of QTextLayouts
versus QTextDocuments when the leading is large enough to impact
the font height. See QTBUG-82954 for details on this.

Since the two were assumed to be the same, this change causes
a test failure with certain fonts. To work around it, we expect
the test failure for the case where the text format is
RichText (and QTextDocument is going to be used internally),
and we detect that the font exhibits the problem. For TextEdits,
QTextDocument is always used, so we also expect failure there.

Note that this is equivalent to the fix for the same issue
in the QLabel test: c4ef0b92d5fb2c621e880347bd48d01b6f31eb24

Fixes: QTBUG-83839
Task-number: QTBUG-82954
Change-Id: Ic32e0d835480b42b68bb452301ec4bfa6ce30d3a
Reviewed-by: Liang Qi <liang.qi@qt.io>