Fix Narrator spelling password characters in Quick TextInput
Quick was not setting the "passwordEdit" accessible state for TextInput
controls set with a password-editing echo mode. Thus the platform
accessibility layer would not let the accessibility client (Microsoft
Narrator, in the reported case) know that the text being edited was a
password, making it spell the letters instead of something like "hidden".

Fixes: QTBUG-82433
Pick-to: 5.15
Change-Id: Ic659f031f6bad2ae3d7e500333dc8e2f27797a08
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>