Remove the 'windows' render loop
For compatibility, QSG_RENDER_LOOP=windows shows a warning and uses
basic.

For OpenGL proper, threaded is the default on Windows since Qt 5.5, so
this will not affect there in any way.

By default the windows loop was only used with ANGLE, or a non-standard opengl32
implementation like the Mesa llvmpipe shipped with the pre-built Qt packages.
(and only on Windows, as the name suggests, even though the implementation of this
loop is not tied to Windows in any way)

For these we default to basic now. This will make no difference in practice:
The windows render loop's value is that it takes the custom animation driver
(advanced per frame, assuming that vsync-based throttling is active) approach from
the threaded loop, while behaving like basic otherwise, thus providing smoother
animations compared to basic, in theory. But due to its relatively little use (not
the default since 5.5), ANGLE being removed in the near future, and not
making much sense for the software rasterizer cases (those will have
bigger problems than smoothly advancing animations, often providing no vsync,
so they are better off with basic anyways), it makes no sense to invest into
adding QRhi support for this loop too.

Task-number: QTBUG-78578
Task-number: QTBUG-82997
Change-Id: I8ccd2e5e375799df5bb5018df247aa5b8197032a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>