Expect failure for certain fonts in TextArea test
Change f761ad3cd9ad1252f24b76ae413298dc7bed8af3 in qtbase exposes
a pre-existing difference in measuring the height of QTextLayouts
versus QTextDocuments when the leading is large enough to impact
the font height. See QTBUG-82954 for details on this.

Note that this is equivalent to the fix for the same issue
in the QLabel test: c4ef0b92d5fb2c621e880347bd48d01b6f31eb24

Task-number: QTBUG-82954
Change-Id: Idee81dcceefe594059a06bd1f090650acd738f79
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>