CMake: Adjust and regenerate tools project
Checking for android_app scope does not make sense in subdirs
projects, it's expected that it is explicitly set as CONFIG +=
android_app in a standalone executable project. When the config
is set, instead of creating an .so file, a true Android CLI
executable is created.

This is seemingly a relic from a time when there existed an
Android mkspec that allowed building Qt without the .so
requirement.

Thus remove the android_app scope.

Also exchange the static scope, with a contains(CONFIG, static)
scope, so that pro2cmake can deal with the condition better.

Keep the qmlcachegen target commented, due to it being built
in the src project.

Finally regenerate the CMake project, to exclude certain apps
being built by default, thus fixing failures when building a
simulator_and_device iOS configuration.

Task-number: QTBUG-80899
Change-Id: Ief89d85c3a238c4a002ab6d98ee2596508cf130b
Reviewed-by: Simon Hausmann <hausmann@gmail.com>