Add WheelHandler.blocking property
Just as PointHandler is meant for providing passive feedback without
interfering with delivery, it should be possible to use WheelHandler
to notice the wheel rotation and let the event continue propagating
to other items and handlers. In that case, set the blocking property
to false, which tells it not to call QEventPoint::setAccepted().

[ChangeLog][QtQuick][WheelHandler] WheelHandler now has a property
called blocking, which is true by default; but if set to false, it
allows wheel events to propagate to items "under" this handler's
parent, and their WheelHandlers.

Change-Id: I44259709a3d8b2cbc1c6c11a88d669d847309924
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>