TableView: during layout, set item width before height
Text items like TextEdit will calculate their implicit height
based on the assigned width (because of word-wrap). It's
therefore better to set the width of delegate items first
during a layout than after, since then we also get the
correct height in such (special) cases. The result is
that a rows height will show all the text in a TextEdit, and
not clip it.

Pick-to: 5.14 5.15
Task-number: QTBUG-84046
Change-Id: I1893187027b45649568347ffc3ed5d4d84beaa95
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>