Pre-check for Metal support to please our CI VMs
An attempt to have a temporary solution for not having Metal support in
the CI VMs. In order to run graphical Qt Quick autotests, we need to
decide upfront if we want to go with Metal or try something else.
Leaving the decision to the point of initializing the scenegraph (and
so creating the QRhi instance) is not acceptable as that is way too
late.

If this works out, we can add a private helper in gui that does a more
lightweight check directly via the Metal APIs. But for now, just create
a QRhi temporarily.

For qtbase autotests (like tst_qrhi) the problem does not apply because
those simply go through all QRhi backends and exercise all that succeed
to initialize, skipping the failing ones. But a real application, and so
Qt Quick, does not work like that.

Task-number: QTBUG-84067
Change-Id: I52f31871eb72c8d57bc604391963f0d022a3321c
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>