Only alter rhi backend on Win7 and Metalless-macOS when not explicitly requested
When forcing a given backend from the C++ API or QSG_RHI_BACKEND, the
user/application's request must be honored. Otherwise confusing
situations arise. If the requested backend fails, let's just fail.

It is the "just pick a platform specific default" case where we have the
flexibility of choosing between the backends.

Change-Id: If30c329bc7ce8ec26abee0ae2d4bcaef53df4718
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>