Have the basic render loop notify when the renderpass desc. disappears
Unlike the threaded one, this runs into the well-known issue of
referencing the QRhiRenderPassDescriptor from data structures that are
tied to the rendercontext.

In the 'basic' loop there is one rendercontext that is reused for all
windows. With threaded there would be one rendercontext per window,
because there each window gets its own render thread, and so the problem
cannot arise.

Fix up the basic loop accordingly.

Change-Id: Iea6713c6167ead7df96ac5f01d7b92d2ec57ce71
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>