Check thread->isRunning before posting WMReleaseSwapchainEvent
If the window becomes "obscured" (hidden), rendering (i.e. the render
thread) stops. Getting a PlatformSurfaceAboutToBeDestroyed afterwards
led to sending (and blocking) the WMReleaseSwapchainEvent to the render
thread. This lead to a deadlock. While this is not something that
happens normally, the problem is showcased by the "headless" case of
the qquickwindow autotest which has a quite aggressive destroy() call
on the QQuickWindow.

Change-Id: I4726b76646564162af645508e08e35e268fea7e4
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>