Allow aborting syncAndRender() when sync results in no changes
This is the behavior with direct OpenGL. With QRhi, we avoided returning
early because QRhi::beginFrame() was already called, so rendering the
frame seemed like a better choice, at first at least. (the SkipPresent
flag was likely introduced at a later point)

Now we just do an endFrame() with SkipPresent set, and then return
without moving on to the rendering phase.

This fixes the noUpdateWhenNothingChanges case in the qquickwindow
autotest.

Additionally, it is suspected that this also fixes the qquicklayout auto
test failures: there we have tests that wait for the frameSwapped()
signal. Until now we did not abort the update when sync resulted in no
changes (with RHI), which confuses said tests because there is an
unexpected frameSwapped() signal emitted.

Change-Id: Id6a859d27a525ed2c16c6b331b1ad5debabfcd77
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>